+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [0.7.4] - 2019-05-15
 
 ### Fixed
 
@@ -198,7 +198,8 @@ This is a major release with some API changes.
 
 
 
-[Unreleased]: https://github.com/sdispater/cleo/compare/0.7.3...master
+[Unreleased]: https://github.com/sdispater/cleo/compare/0.7.4...master
+[0.7.4]: https://github.com/sdispater/cleo/releases/tag/0.7.4
 [0.7.3]: https://github.com/sdispater/cleo/releases/tag/0.7.3
 [0.7.2]: https://github.com/sdispater/cleo/releases/tag/0.7.2
 [0.7.1]: https://github.com/sdispater/cleo/releases/tag/0.7.1